refactor: split messages + specialist detail into section components
